Separate CSV and Google Sheet loader
